requirements.txt
altair==5.3.0
annotated-types==0.7.0
attrs==24.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
gitpython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
idna==3.7
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
